requirements.txt
# requirements.txt
Flask==3.0.2
tensorflow==2.18.0
numpy==1.26.4
pillow==12.0.0
Werkzeug==3.0.1
flask-cors==4.0.0
groq
python-dotenv
